+++ ChangeLog.md
@@ -8,6 +8,7 @@
 
     - configure() now warns for unmanaged keys
     - Use pwix:logger universal logger, thus bumping minor candidate version number
+    - Def.toTabular() now provides the name to the column if it is set
 
 ### 1.6.0